Fix Windows build errors. (#8381)
* Adapt toolchain file to UE versions other than 5.3.

* More 5.5-related fixes.

* Remove unused SetFolderName call and fix GetRenderQuery call.

* Fix several deprecation warnings.

* More deprecation fixes.

* Upgrade enum to string code.

* Add a check for null weather.

* Update readme with UE5.5

* Update unreal-side includes, add include guards.

* Remove unwanted changes from merge.

* More fixes.

* Fix non-unity build error.

* Uncomment code and remove unwanted files (these are generated at CMake configure time).

* Fix Ubuntu build.

* Add missing initializers.

* Fix windows minor build issue.

---------

+++ Unreal/CarlaUnreal/Plugins/Carla/Source/Carla/Sensor/ImageUtil.cpp
@@ -7,10 +7,13 @@
 #include <Carla/Sensor/ImageUtil.h>
 #include <Carla/Sensor/ShaderBasedSensor.h>
 #include <Carla/Carla.h>
-#include <Runtime/RHI/Public/RHISurfaceDataConversion.h>
-#include <Runtime/ImageWriteQueue/Public/ImageWriteQueue.h>
+
+#include <util/ue-header-guard-begin.h>
+#include <RHISurfaceDataConversion.h>
+#include <ImageWriteQueue.h>
 #include <HighResScreenshot.h>
 #include <RHIGPUReadback.h>
+#include <util/ue-header-guard-end.h>
 
 
 
@@ -101,11 +104,11 @@ namespace ImageUtil
       ConvertRawR24G8DataToFColor(Extent.X, Extent.Y, (uint8*)PixelData, SourcePitch, Out.GetData(), Flags);
       break;
     case PF_R32_FLOAT: // Depth
-		  for (uint32 Y = 0; Y < Extent.Y; Y++)
+		  for (uint32 Y = 0; Y < (uint32)Extent.Y; Y++)
 		  {
 		  	auto SrcPtr = (float*)((uint8*)PixelData + Y * SourcePitch);
 		  	auto DestPtr = Out.GetData() + Y * Extent.X;
-		  	for (uint32 X = 0; X < Extent.X; X++)
+		  	for (uint32 X = 0; X < (uint32)Extent.X; X++)
 		  	{
 		  		*DestPtr = FLinearColor(SrcPtr[0], 0.f, 0.f, 1.f).QuantizeRound();
 		  		++SrcPtr;
